backend/pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "RAG pipeline for URL ingestion, embedding generation, and vector storage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.0.0",
    "qdrant-client>=1.9.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.12.0",
    "openai-agents>=0.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.5",
    "sse-starlette>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]
